requirements.txt
scrapy
paho-mqtt
